+++ delivery_myparcel/models/delivery_carriers/delivery_myparcel_dhl_europlus.py
@@ -14,8 +14,6 @@ from odoo.exceptions import ValidationError
 
 _logger = logging.getLogger()
 
-MYPARCEL_BASE_TRACK_TRACE_URL = 'https://myparcel.me/track-trace'
-
 
 class ProviderMyparcelDHLEuroplus(models.Model):
     _inherit = ['delivery.carrier', 'myparcel.mixin']
@@ -49,6 +47,7 @@ class ProviderMyparcelDHLEuroplus(models.Model):
             "package_type": self._get_package_type(self),
             "delivery_type": self._get_delivery_code(self.delivery_type),
             "label_description": order.name,
+            "weight": order.shipping_weight,
             "saturday_delivery": 0,
         }
 
@@ -58,6 +57,9 @@ class ProviderMyparcelDHLEuroplus(models.Model):
         options.update(self.generate_custom_field_options(fields=[
             ("signature", "x_aa_mp_signing"),
             ("hide_sender", "x_aa_mp_hide_sender"),
+            # ("return", "x_aa_mp_direct_return"),
+            ("insurance", "x_aa_mp_insurance"),
+            ("insurance_price", "x_aa_mp_insurance_pricelist_id"),
         ], order=order, carrier=self))
         return options
 
@@ -78,7 +80,7 @@ class ProviderMyparcelDHLEuroplus(models.Model):
         return BaseProviderMyParcel.base_myparcel_get_label(self, picking=picking)
 
     def myparcel_dhl_europlus_get_tracking_link(self, picking):
-        track_trace_base = MYPARCEL_BASE_TRACK_TRACE_URL
+        track_trace_base = BaseProviderMyParcel.get_myparcel_base_tracking_url()
         url = f'{track_trace_base}/{picking.carrier_tracking_ref}/{picking.partner_id.zip}/{picking.partner_id.country_id.code}'
         # url = f"https://my.dhlparcel.nl/home/tracktrace/{picking.carrier_tracking_ref}/{picking.partner_id.zip}"
         return url