+++ delivery_myparcel/models/myparcel_request.py
@@ -14,7 +14,6 @@ import logging
 import requests
 import json
 
-
 _logger = logging.getLogger(__name__)
 
 MYPARCEL_BASE_API_URL = 'https://api.myparcel.nl'
@@ -104,58 +103,172 @@ class MyParcelRequest:
             if recipient.country_id:
                 recipient_dict['cc'] = recipient.country_id.code
             else:
-                raise ValidationError(_('Country code is required for creating a shipment. Please check the delivery address.'))
+                raise ValidationError(
+                    _('Country code is required for creating a shipment. Please check the delivery address.'))
             if recipient.state_id:
-                recipient_dict['region'] = recipient.state_id.name
+                if recipient.country_id.code == 'US':
+                    recipient_dict['state'] = recipient.state_id.code
+                else:
+                    recipient_dict['region'] = recipient.state_id.name
             else:
-                raise ValidationError(_('State is required for creating a shipment. Please check the delivery address.'))
+                raise ValidationError(
+                    _('State is required for creating a shipment. Please check the delivery address.'))
+
             if recipient.city:
                 recipient_dict['city'] = recipient.city
             else:
                 raise ValidationError(_('City is required for creating a shipment. Please check the delivery address.'))
 
-            # get street number
-            if recipient.street2:
-                recipient_dict['street'] = recipient.street
-                recipient_dict['number'] = recipient.street2
+            # get street number for EU countries
+            if recipient.country_id.x_aa_mp_is_european:
+                if recipient.street2:
+                    recipient_dict['street'] = recipient.street
+                    recipient_dict['number'] = recipient.street2
+                else:
+                    try:
+                        if re.match(r"^(.*?)(\d+)$", recipient.street.strip()):
+                            street_number = re.match(r"^(.*?)(\d+)$", recipient.street.strip()).group(2).strip()
+                            recipient_dict['street'] = recipient.street.replace(str(street_number), '')
+                            recipient_dict['number'] = street_number
+                            _logger.warning(F'number {street_number}')
+                        else:
+                            raise ValidationError(
+                                _('Street number not found in street field. Please check the delivery address.'))
+                    except AttributeError:
+                        raise ValidationError(
+                            _('Street number not found in street field. Please check the delivery address.'))
+
             else:
-                try:
-                    if re.match(r"^(.*?)(\d+)$", recipient.street.strip()):
-                        street_number = re.match(r"^(.*?)(\d+)$", recipient.street.strip()).group(2).strip()
-                        recipient_dict['street'] = recipient.street.replace(str(street_number), '')
-                        recipient_dict['number'] = street_number
-                        _logger.warning(F'number {street_number}')
-                    else:
-                        raise ValidationError(_('Street number not found in street field. Please check the delivery address.'))
-                except AttributeError:
-                    raise ValidationError(_('Street number not found in street field. Please check the delivery address.'))
+                if recipient.street:
+                    recipient_dict['street'] = recipient.street
+                    recipient_dict['number'] = ''
+                    recipient_dict['street_additional_info'] = recipient.street2
+                else:
+                    raise ValidationError(
+                        _('Complete address is required for creating a shipment. Please check the delivery address.'))
 
             if recipient.zip:
                 recipient_dict['postal_code'] = recipient.zip
             else:
-                raise ValidationError(_('Postal code is required for creating a shipment. Please check the delivery address.'))
+                raise ValidationError(
+                    _('Postal code is required for creating a shipment. Please check the delivery address.'))
             if recipient.name:
                 recipient_dict['person'] = recipient.name
                 _logger.warning(F'recipient.name {recipient.name}')
             else:
-                raise ValidationError(_('Recipient name is required for creating a shipment. Please check the delivery address.'))
+                raise ValidationError(
+                    _('Recipient name is required for creating a shipment. Please check the delivery address.'))
 
             if carrier:
                 if carrier.x_aa_mp_phone_send_with_shipment and recipient.phone:
                     recipient_dict['phone'] = recipient.phone
                 elif carrier.x_aa_mp_phone_send_with_shipment and not recipient.phone:
-                    raise ValidationError(_('Phone number is required for creating a shipment. Please check the delivery address.'))
+                    raise ValidationError(
+                        _('Phone number is required for creating a shipment. Please check the delivery address.'))
                 if carrier.x_aa_mp_email_send_with_shipment and recipient.email:
                     if re.match(r"^[^@]+@[^@]+\.[^@]+$", recipient.email):
                         recipient_dict['email'] = recipient.email
                     else:
                         raise ValidationError(_('Invalid email address provided. Please check the delivery address.'))
                 elif carrier.x_aa_mp_email_send_with_shipment and not recipient.email:
-                    raise ValidationError(_('Email address is required for creating a shipment. Please check the delivery address.'))
+                    raise ValidationError(
+                        _('Email address is required for creating a shipment. Please check the delivery address.'))
 
             return recipient_dict
         else:
-            raise ValidationError(_('Correct contact information is required for creating a shipment. Please check the delivery address.'))
+            raise ValidationError(
+                _('Correct contact information is required for creating a shipment. Please check the delivery address.'))
+
+    def _get_customs_declaration(self, record, carrier_code, recipient):
+        _logger.warning(F'record._name {record._name}')
+        customs_declaration = {}
+        # TODO: Put first part ou of loop depending on if can be done
+        if record.sudo()._name == 'sale.order':
+            customs_declaration['contents'] = 1
+            # TODO: implement package contents (commercial goods, commercial samples etc...)
+            customs_declaration['invoice'] = record.name  # TODO: implement invoice number
+            customs_declaration['weight'] = int(
+                record.shipping_weight * 1000) if record.shipping_weight else ValidationError(
+                _('No shipping weight found. Please check if the product is defined correctly.'))
+            items_list = []
+            for order_line in record.order_line:
+                if order_line.product_id.type != 'service':
+                    # TODO: CHECK hs code or ISIC code and check country_of_origin model
+                    # TODO: get classification code based on carrier
+                    # Get list for each carrier required HS code length
+                    # Check number of characters per HS code
+                    if order_line.product_id.hs_code:
+                        if (len(order_line.product_id.hs_code) == 10 and recipient.country_id.cc == 'US') or (
+                                len(order_line.product_id.hs_code) >= 6 and recipient.country_id.cc != 'US'):
+                            classification = order_line.product_id.hs_code
+                        else:
+                            raise ValidationError(_('The HS Code does not have the correct amount of digits (minimum 6 '
+                                                    'digits and minimum 10 digits for US shipments.'))
+                    else:
+                        raise ValidationError(
+                            _('No HS Code found for this product. Please check if the product is defined correctly.'))
+
+                    items_list.append({
+                        'description': order_line.name,
+                        'amount': int(order_line.product_uom_qty),
+                        'weight': int(
+                            order_line.product_id.weight * 1000) if order_line.product_id.weight else ValidationError(
+                            _('No weight found for the product. Please check if the product is defined correctly.')),
+                        'item_value': {
+                            "amount": int(order_line.price_unit * order_line.product_uom_qty * 100),
+                            "currency": record.currency_id.name or "EUR",
+                        },
+                        'classification': classification,
+                        'country': order_line.product_id.country_of_origin.code if order_line.product_id.country_of_origin else ValidationError(
+                            _('No country of origin found for this product. Please check if the product is defined correctly.')),
+                    })
+
+            _logger.warning(F'items for douane declaration {items_list}')
+            customs_declaration['items'] = items_list
+
+        elif record.sudo()._name == 'stock.picking':
+            # TODO: Put first part ou of loop depending on if can be done
+            customs_declaration['contents'] = 1
+            # TODO: implement package contents (commercial goods, commercial samples etc...)
+            customs_declaration['invoice'] = record.name  # TODO: implement invoice number
+            customs_declaration['weight'] = int(
+                record.shipping_weight * 1000) if record.shipping_weight else ValidationError(
+                _('No shipping weight found. Please check if the product is defined correctly.'))
+            items_list = []
+            for stock_move in record.move_ids:
+                order_line = record.sale_id.order_line.search(
+                    [('product_template_id', '=', stock_move.product_id.product_tmpl_id.id)], limit=1)
+                # TODO: CHECK hs code or ISIC code and check country_of_origin model
+                # Check number of characters per HS code
+                if order_line.product_id.hs_code:
+                    if (len(order_line.product_id.hs_code) == 10 and recipient.country_id.cc == 'US') or (
+                            len(order_line.product_id.hs_code) in [6, 8, 10]):
+                        classification = order_line.product_id.hs_code
+                    else:
+                        raise ValidationError(_('The HS Code does not have the correct amount of digits.'))
+                else:
+                    raise ValidationError(
+                        _('No HS Code found for this product. Please check if the product is defined correctly.'))
+
+                items_list.append({
+                    'description': stock_move.product_id.name,
+                    'amount': int(stock_move.quantity),
+                    'weight': int(
+                        stock_move.product_id.weight * 1000) if stock_move.product_id.weight else ValidationError(
+                        _('No weight found for the product. Please check if the product is defined correctly.')),
+                    'item_value': {
+                        "amount": int(order_line.price_unit * order_line.product_uom_qty * 100),
+                        "currency": record.currency_id.name or "EUR",
+                    },
+                    'classification': classification,
+                    'country': stock_move.product_id.country_of_origin.code if stock_move.product_id.country_of_origin else ValidationError(
+                        _('No country of origin found for this product. Please check if the product is defined correctly.')),
+                })
+
+            _logger.warning(F'items for douane declaration {items_list}')
+            customs_declaration['items'] = items_list
+
+        return customs_declaration
 
     def post_shipment_body(self, recipient, options, record=None, carrier_code='', carrier=None):
         # Make recipient
@@ -171,24 +284,27 @@ class MyParcelRequest:
         # TODO: options implementeren
 
         # Add insurance
-        if record and options:
-            if options['insurance'] == 1:
-                ins_price = options['insurance_price']
-                _logger.warning(F'insurance price in request {ins_price}')
-                insured_amount = record.x_aa_mp_insurance_pricelist_id.search([('id', '=', options['insurance_price'].id)], limit=1)
-                if insured_amount:
-                    insured_amount = int(insured_amount.x_aa_mp_insurance_selection)
-                options['insurance'] = {
-                    "amount": insured_amount,
-                    "currency": record.currency_id.name or "EUR"
-                }
-                options.pop('insurance_price')
+        _logger.warning(F'options in post_shipment_body {options}')
+        if options and 'insurance' in options:
+            if record:
+                if options['insurance'] == 1:
+                    ins_price = options['insurance_price']
+                    _logger.warning(F'insurance price in request {ins_price}')
+                    insured_amount = record.x_aa_mp_insurance_pricelist_id.search(
+                        [('id', '=', options['insurance_price'].id)], limit=1)
+                    if insured_amount:
+                        insured_amount = int(insured_amount.x_aa_mp_insurance_selection)
+                    options['insurance'] = {
+                        "amount": insured_amount,
+                        "currency": record.currency_id.name or "EUR"
+                    }
+                    options.pop('insurance_price')
+                else:
+                    options.pop('insurance')
+                    options.pop('insurance_price')
             else:
                 options.pop('insurance')
                 options.pop('insurance_price')
-        else:
-            options.pop('insurance')
-            options.pop('insurance_price')
 
         _logger.warning(F'options in body construction {options}')
 
@@ -199,6 +315,18 @@ class MyParcelRequest:
             "carrier": carrier_code,
         }
 
+        # Shipping weight
+        # TODO: Add height, length, width of product (Discuss if necessary)
+        if options and 'weight' in options:
+            if options['weight'] == 0:
+                options['weight'] = 0.001  # MyParcel requires a minimum weight of 1 gram
+            shipping_weight = {
+                "weight": int(options['weight']) * 1000,
+            }
+            options.pop('weight')
+            shipments['physical_properties'] = shipping_weight
+
+        # Multi-collo shipments
         if carrier_code == 1 and record:
             if record._name == 'stock.picking' and record.package_level_ids:
                 if len(record.package_level_ids) > 1:
@@ -221,6 +349,11 @@ class MyParcelRequest:
                                 'package_name': package.package_id.name,
                             })
                         })
+
+        # Douane information
+        if not recipient.country_id.x_aa_mp_is_european:
+            customs_declaration = self._get_customs_declaration(record, carrier_code, recipient)
+            shipments['customs_declaration'] = customs_declaration
 
         data = {
             'data': {