+++ delivery_myparcel/wizard/choose_delivery_carrier.py
@@ -20,6 +20,8 @@ class ChooseDeliveryCarrier(models.TransientModel):
 
     x_aa_mp_insurance_pricelist_id = fields.Many2one('myparcel.insurance.price', string='Price')
 
+    commitment_date = fields.Datetime(related='order_id.commitment_date', string='Delivery Date', readonly=False)
+
     @api.onchange('carrier_id')
     def _onchange_carrier_id(self):
         if self.order_id and self.order_id.carrier_id and self.order_id.carrier_id == self.carrier_id:
@@ -28,8 +30,8 @@ class ChooseDeliveryCarrier(models.TransientModel):
                 'x_aa_mp_signing': self.order_id.x_aa_mp_signing,
                 'x_aa_mp_only_receiver': self.order_id.x_aa_mp_only_receiver,
                 'x_aa_mp_receiving_code': self.order_id.x_aa_mp_receiving_code,
-                # 'x_aa_mp_direct_return': self.order_id.x_aa_mp_direct_return,
-                # 'x_aa_mp_home_delivery': self.order_id.x_aa_mp_home_delivery,
+                'x_aa_mp_direct_return': self.order_id.x_aa_mp_direct_return,
+                'x_aa_mp_large_package': self.order_id.x_aa_mp_large_package,
                 'x_aa_mp_hide_sender': self.order_id.x_aa_mp_hide_sender,
                 'x_aa_mp_insurance': self.order_id.x_aa_mp_insurance,
                 'x_aa_mp_insurance_pricelist_id': self.order_id.x_aa_mp_insurance_pricelist_id,
@@ -41,8 +43,8 @@ class ChooseDeliveryCarrier(models.TransientModel):
                 'x_aa_mp_signing': self.carrier_id.x_aa_mp_signing,
                 'x_aa_mp_only_receiver': self.carrier_id.x_aa_mp_only_receiver,
                 'x_aa_mp_receiving_code': self.carrier_id.x_aa_mp_receiving_code,
-                # 'x_aa_mp_direct_return': self.carrier_id.x_aa_mp_direct_return,
-                # 'x_aa_mp_home_delivery': self.carrier_id.x_aa_mp_home_delivery,
+                'x_aa_mp_direct_return': self.carrier_id.x_aa_mp_direct_return,
+                'x_aa_mp_large_package': self.carrier_id.x_aa_mp_large_package,
                 'x_aa_mp_hide_sender': self.carrier_id.x_aa_mp_hide_sender,
                 'x_aa_mp_insurance': False,
                 'x_aa_mp_insurance_pricelist_id': False,
@@ -52,11 +54,10 @@ class ChooseDeliveryCarrier(models.TransientModel):
             self.write({
                 'x_aa_mp_age_control': False,
                 'x_aa_mp_signing': False,
-                # 'x_aa_mp_insurance': False,
                 'x_aa_mp_only_receiver': False,
                 'x_aa_mp_receiving_code': False,
-                # 'x_aa_mp_direct_return': False,
-                # 'x_aa_mp_home_delivery': False,
+                'x_aa_mp_direct_return': False,
+                'x_aa_mp_large_package': False,
                 'x_aa_mp_insurance': False,
                 'x_aa_mp_insurance_pricelist_id': False,
                 'x_aa_mp_hide_sender': False,
@@ -69,6 +70,8 @@ class ChooseDeliveryCarrier(models.TransientModel):
         'x_aa_mp_signing',
         'x_aa_mp_only_receiver',
         'x_aa_mp_receiving_code',
+        'x_aa_mp_large_package',
+        'x_aa_mp_direct_return',
         'x_aa_mp_hide_sender',
         'x_aa_mp_insurance',
         'x_aa_mp_insurance_pricelist_id'
@@ -94,21 +97,19 @@ class ChooseDeliveryCarrier(models.TransientModel):
                         self.carrier_message = _('Same day delivery is not available for this delivery.')
                 if 'dhl_europlus' in self.carrier_id.delivery_type:
                     self.carrier_message = _('Signature is always required for Europlus deliveries.')
+                if 'gls' in self.carrier_id.delivery_type:
+                    self.carrier_message = _('Signature is always required outside the Netherlands for GLS deliveries.')
 
     def _get_delivery_rate(self):
-        # self_context = self.with_context(x_aa_mp_age_control=self.x_aa_mp_age_control,
-        #                                  x_aa_mp_signing=self.x_aa_mp_signing,
-        #                                  x_aa_mp_only_receiver=self.x_aa_mp_only_receiver,
-        #                                  x_aa_mp_direct_return=self.x_aa_mp_direct_return,
-        #                                  x_aa_mp_home_delivery=self.x_aa_mp_home_delivery,
-        #                                  x_aa_mp_hide_sender=self.x_aa_mp_hide_sender)
         self_context = self.with_context(x_aa_mp_age_control=self.x_aa_mp_age_control,
                                          x_aa_mp_signing=self.x_aa_mp_signing,
                                          x_aa_mp_only_receiver=self.x_aa_mp_only_receiver,
                                          x_aa_mp_receiving_code=self.x_aa_mp_receiving_code,
+                                         x_aa_mp_large_package=self.x_aa_mp_large_package,
+                                         x_aa_mp_direct_return=self.x_aa_mp_direct_return,
                                          x_aa_mp_hide_sender=self.x_aa_mp_hide_sender,
                                          x_aa_mp_insurance=self.x_aa_mp_insurance,
-                                         x_aa_mp_insurance_pricelist_id=self.x_aa_mp_insurance_pricelist_id,#.id if self.x_aa_mp_insurance_pricelist_id else False,
+                                         x_aa_mp_insurance_pricelist_id=self.x_aa_mp_insurance_pricelist_id,
                                          x_aa_mp_allow_sameday_delivery=self.x_aa_mp_allow_sameday_delivery)
         return super(ChooseDeliveryCarrier, self_context)._get_delivery_rate()