+++ delivery_myparcel/CHANGELOG.md
@@ -8,21 +8,36 @@ All notable changes to this project will be documented in this file.
 
 * The module is currently still in development. New features/carriers are actively being implemented.
 
+## 2025-08-20 (18.0.0.3.0)
+
+### Added
+
+* Integration for MyParcel carrier: DPD
+* Integration for MyParcel carrier: UPS Standard/Express Saver
+* Customs information for international (non-EU) shipments
+
+### Updated
+
+* Large package option for the Post NL carrier
+* Direct return option for the Post NL carrier
+* Shipping weight always added to shipment
+
 ## 2025-07-18 (18.0.0.2.0)
 
 ### Added
 
-* Integration for MyParcel carrier: PostNL
+* Integration for MyParcel carrier: Post Nl
 * Integration for MyParcel carrier: DHL Parcel Connect
-* Multi-Collo package for PostNL
+* Multi-Collo package for Post NL
 * Ability to select package types per carrier and select them in picking 'Put in Pack' while showing only available 
   options depending on the carrier
 * Added insurance option for MyParcel shipments with ability to set a fixed or dynamic price for insurance
 
 ### Updated
+
 * Contact information check for MyParcel shipments
-* Access Denied error handling (using incorrect MyParcel API key)
-* Use MyParcel's custom Track & Trace link instead of carrier specific url
+* Access Denied error handling
+* Use MyParcel's custom track trace link instead of carrier url
 
 ## 2025-06-25 (18.0.0.1.2)